requirements.txt
# UniProtKB Protein Finder Requirements

# Core dependencies
requests>=2.28.0          # HTTP requests to UniProtKB REST API

# Optional dependencies (for enhanced functionality)
python-dotenv>=0.19.0    # Load environment variables from .env file

# The application uses Python's built-in modules for:
# - Tkinter (GUI)
# - threading (responsive UI)
# - json (data parsing & export)
# - csv (search history logging)
# - pathlib (cross-platform file handling)
